Config 분리& useCheckBox 옵션 추가
